backend/backend_fetch/requirements.txt
functions-framework==3.*
flask==2.1.0
google-cloud-error-reporting==1.6.0
google-cloud-firestore==2.7.2


